rearrange components and types to facilitate testing with exporting types
