+++ mobile/app/(tabs)/explore.tsx
@@ -18,17 +18,18 @@ import { Ionicons } from '@expo/vector-icons';
 import { ThemedText } from '@/components/ThemedText';
 import { useNavigation, useFocusEffect } from '@react-navigation/native';
 import { AuthContext } from '../_layout';
+import PostItem from '../components/PostItem';
 
 const HOST = Platform.select({ android: '10.0.2.2', ios: 'localhost', web: 'localhost' });
 const API_BASE = `http://${HOST}:8080`;
 
-// Frontend Comment Type
-type CommentData = {
+// Define the CommentData interface
+interface CommentData {
   commentId: number;
-  content: string;
-  createdAt: string;
   username: string;
-};
+  content: string;
+  createdAt: string | Date; // Can be string from API, then converted to Date
+}
 
 type Post = {
   id: number;
@@ -39,312 +40,6 @@ type Post = {
   photoUrl: string | null;
   likedByUser: boolean;
 };
-
-// --- CommentItemDisplay Component ---
-interface CommentItemDisplayProps {
-  comment: CommentData;
-  commentTextColor: string;
-  commentUsernameColor: string;
-  commentBorderColor: string;
-  loggedInUsername: string | null;
-  onDeleteComment: (commentId: number) => void;
-  deleteIconColor: string;
-  // --- NEW PROPS for edit functionality ---
-  editIconColor: string;
-  onTriggerEdit: (comment: CommentData) => void;
-  isEditingThisComment: boolean;
-  editedContent: string;
-  onEditContentChange: (newText: string) => void;
-  onSaveEditedComment: (commentId: number) => void;
-  onCancelEdit: () => void;
-  isSavingEdit: boolean;
-  // --- END NEW PROPS for edit ---
-}
-
-function CommentItemDisplay({
-  comment,
-  commentTextColor,
-  commentUsernameColor,
-  commentBorderColor,
-  loggedInUsername,
-  onDeleteComment,
-  deleteIconColor,
-  // --- NEW PROPS for edit functionality ---
-  editIconColor,
-  onTriggerEdit,
-  isEditingThisComment,
-  editedContent,
-  onEditContentChange,
-  onSaveEditedComment,
-  onCancelEdit,
-  isSavingEdit,
-  // --- END NEW PROPS for edit ---
-}: CommentItemDisplayProps) {
-  const isOwner = loggedInUsername && comment.username === loggedInUsername;
-  const colorScheme = useColorScheme(); // For save/cancel button text color
-
-  if (isOwner && isEditingThisComment) {
-    return (
-      <View style={[styles.commentItemContainer, { borderBottomColor: commentBorderColor }]}>
-        <View style={styles.commentHeader}>
-          <ThemedText style={[styles.commentUsername, { color: commentUsernameColor }]}>{comment.username} (editing)</ThemedText>
-        </View>
-        <TextInput
-          style={[styles.commentEditInput, { borderColor: editIconColor, color: commentTextColor, backgroundColor: colorScheme === 'dark' ? '#2C2C2E' : '#F9F9F9' }]}
-          value={editedContent}
-          onChangeText={onEditContentChange}
-          multiline
-          autoFocus
-          editable={!isSavingEdit}
-        />
-        <View style={styles.editActionsContainer}>
-          <TouchableOpacity
-            style={[styles.editActionButton, { backgroundColor: '#888' }]}
-            onPress={onCancelEdit}
-            disabled={isSavingEdit}
-          >
-            <ThemedText style={styles.editActionButtonText}>Cancel</ThemedText>
-          </TouchableOpacity>
-          <TouchableOpacity
-            style={[styles.editActionButton, { backgroundColor: editIconColor }]}
-            onPress={() => onSaveEditedComment(comment.commentId)}
-            disabled={isSavingEdit || !editedContent.trim()}
-          >
-            {isSavingEdit ? (
-              <ActivityIndicator size="small" color="#FFFFFF" />
-            ) : (
-              <ThemedText style={styles.editActionButtonText}>Save</ThemedText>
-            )}
-          </TouchableOpacity>
-        </View>
-      </View>
-    );
-  }
-
-  return (
-    <View style={[styles.commentItemContainer, { borderBottomColor: commentBorderColor }]}>
-      <View style={styles.commentHeader}>
-        <ThemedText style={[styles.commentUsername, { color: commentUsernameColor }]}>{comment.username}</ThemedText>
-        {isOwner && (
-          <View style={styles.commentOwnerActions}>
-            <TouchableOpacity onPress={() => onTriggerEdit(comment)} style={styles.commentActionButton}>
-              <Ionicons name="pencil-outline" size={18} color={editIconColor} />
-            </TouchableOpacity>
-            <TouchableOpacity onPress={() => onDeleteComment(comment.commentId)} style={styles.commentActionButton}>
-              <Ionicons name="trash-outline" size={18} color={deleteIconColor} />
-            </TouchableOpacity>
-          </View>
-        )}
-      </View>
-      <ThemedText style={[styles.commentContent, { color: commentTextColor }]}>{comment.content}</ThemedText>
-      <ThemedText style={[styles.commentTimestamp, { color: commentTextColor }]}>
-        {new Date(comment.createdAt).toLocaleDateString()}
-      </ThemedText>
-    </View>
-  );
-}
-
-
-// --- PostItem Component ---
-interface PostItemProps {
-  post: Post;
-  cardBackgroundColor: string;
-  iconColor: string;
-  textColor: string;
-  commentInputBorderColor: string;
-  commentInputTextColor: string;
-  commentInputPlaceholderColor: string;
-  commentInputBackgroundColor: string;
-  onLikePress: (postId: number, currentlyLiked: boolean) => void;
-  userType: string | null;
-  loggedInUsername: string | null;
-
-  isExpanded: boolean;
-  commentsList: CommentData[];
-  isLoadingComments: boolean;
-  commentInputText: string; // For new comments
-  isPostingComment: boolean; // For new comments
-  onToggleComments: () => void;
-  onCommentInputChange: (text: string) => void; // For new comments
-  onPostComment: () => void; // For new comments
-  onDeleteComment: (postId: number, commentId: number) => void;
-
-  // --- NEW PROPS for edit functionality ---
-  onTriggerEditComment: (postId: number, comment: CommentData) => void;
-  editingCommentDetailsForPost: { commentId: number; currentText: string; } | null; // Details for *this specific post*
-  onEditCommentContentChange: (newText: string) => void; // Generic handler, ExploreScreen manages which comment text
-  onSaveEditedCommentForPost: (postId: number, commentId: number) => void;
-  onCancelCommentEdit: () => void;
-  isSubmittingCommentEditForPost: boolean; // Specific to this post
-  // --- END NEW PROPS for edit ---
-}
-
-function PostItem({
-  post,
-  cardBackgroundColor,
-  iconColor,
-  textColor,
-  commentInputBorderColor,
-  commentInputTextColor,
-  commentInputPlaceholderColor,
-  commentInputBackgroundColor,
-  onLikePress,
-  userType,
-  loggedInUsername,
-  isExpanded,
-  commentsList,
-  isLoadingComments,
-  commentInputText,
-  isPostingComment,
-  onToggleComments,
-  onCommentInputChange,
-  onPostComment,
-  onDeleteComment,
-  // --- NEW PROPS for edit functionality ---
-  onTriggerEditComment,
-  editingCommentDetailsForPost,
-  onEditCommentContentChange,
-  onSaveEditedCommentForPost,
-  onCancelCommentEdit,
-  isSubmittingCommentEditForPost,
-  // --- END NEW PROPS for edit ---
-}: PostItemProps) {
-
-  const colorScheme = useColorScheme();
-  const commentItemBorderColor = colorScheme === 'dark' ? '#3A3A3C' : '#EAEAEA';
-  const commentUsernameActualColor = colorScheme === 'dark' ? '#E0E0E0' : '#333333';
-  const commentContentActualColor = textColor;
-  const deleteIconActualColor = colorScheme === 'dark' ? '#FF8A80' : '#D9534F';
-  // --- NEW: Define edit icon color ---
-  const editIconActualColor = colorScheme === 'dark' ? '#82B1FF' : '#007AFF'; // A themable blue
-  // --- END NEW ---
-
-  const handleLike = () => {
-    if (userType === 'guest') {
-      Alert.alert("Login Required", "Please log in to like posts.");
-      return;
-    }
-    onLikePress(post.id, post.likedByUser);
-  };
-
-  const canPostComment = userType !== 'guest' && loggedInUsername && !editingCommentDetailsForPost; // Disable new comment if editing one in this post
-
-  return (
-    <View style={[styles.postContainer, { backgroundColor: cardBackgroundColor }]}>
-      {/* ... Post title, image, content, footer (likes/comment count) ... no changes here */}
-      <ThemedText type="title" style={styles.postTitle}>
-        {post.title}
-      </ThemedText>
-      {post.photoUrl && (
-        <Image
-          source={{
-            uri: post.photoUrl.startsWith('http')
-              ? post.photoUrl
-              : `${API_BASE}${post.photoUrl}`,
-          }}
-          style={styles.postImage}
-          onError={(e) => console.warn('Explore: Image failed to load:', e.nativeEvent.error, post.photoUrl)}
-        />
-      )}
-      <ThemedText style={[styles.postContent, {color: textColor}]} numberOfLines={post.photoUrl ? 2 : 5}>
-        {post.content}
-      </ThemedText>
-
-      <View style={styles.postFooter}>
-        <TouchableOpacity onPress={handleLike} style={styles.footerAction}>
-          <Ionicons
-            name={post.likedByUser ? "heart" : "heart-outline"}
-            size={20}
-            color={post.likedByUser ? 'red' : iconColor}
-          />
-          <ThemedText style={[styles.footerText, { color: post.likedByUser ? 'red' : iconColor, marginLeft: 4 }]}>
-            {post.likes}
-          </ThemedText>
-        </TouchableOpacity>
-
-        <TouchableOpacity onPress={onToggleComments} style={[styles.footerAction, { marginLeft: 16 }]}>
-          <Ionicons name="chatbubble-outline" size={20} color={iconColor} />
-          <ThemedText style={[styles.footerText, { color: iconColor, marginLeft: 4 }]}>
-            {post.comments}
-          </ThemedText>
-        </TouchableOpacity>
-      </View>
-
-
-      {isExpanded && (
-        <View style={styles.commentsSection}>
-          {isLoadingComments ? (
-            <ActivityIndicator style={{ marginVertical: 15 }} color={iconColor} />
-          ) : commentsList.length === 0 && !editingCommentDetailsForPost ? ( // Check editing too
-            <ThemedText style={[styles.noCommentsText, {color: textColor}]}>No comments yet. Be the first!</ThemedText>
-          ) : (
-            <View style={styles.commentsListContainer}>
-              {commentsList.map(comment => {
-                const isEditingThisComment = editingCommentDetailsForPost?.commentId === comment.commentId;
-                return (
-                  <CommentItemDisplay
-                    key={comment.commentId}
-                    comment={comment}
-                    commentTextColor={commentContentActualColor}
-                    commentUsernameColor={commentUsernameActualColor}
-                    commentBorderColor={commentItemBorderColor}
-                    loggedInUsername={loggedInUsername}
-                    onDeleteComment={(commentIdToDelete) => onDeleteComment(post.id, commentIdToDelete)}
-                    deleteIconColor={deleteIconActualColor}
-                    // --- Pass props for edit functionality ---
-                    editIconColor={editIconActualColor}
-                    onTriggerEdit={(commentToEdit) => onTriggerEditComment(post.id, commentToEdit)}
-                    isEditingThisComment={isEditingThisComment}
-                    editedContent={isEditingThisComment ? (editingCommentDetailsForPost?.currentText || '') : ''}
-                    onEditContentChange={onEditCommentContentChange}
-                    onSaveEditedComment={() => onSaveEditedCommentForPost(post.id, comment.commentId)}
-                    onCancelEdit={onCancelCommentEdit}
-                    isSavingEdit={isEditingThisComment && isSubmittingCommentEditForPost}
-                    // --- END Pass props for edit ---
-                  />
-                );
-              })}
-            </View>
-          )}
-
-          {/* Add Comment Input Area - only if not guest AND not currently editing a comment in this post */}
-          {canPostComment && (
-            <View style={[styles.addCommentContainer, { borderTopColor: commentItemBorderColor }]}>
-              <TextInput
-                style={[
-                  styles.commentInput,
-                  {
-                    borderColor: commentInputBorderColor,
-                    color: commentInputTextColor,
-                    backgroundColor: commentInputBackgroundColor,
-                  }
-                ]}
-                placeholder="Add a comment..."
-                placeholderTextColor={commentInputPlaceholderColor}
-                value={commentInputText}
-                onChangeText={onCommentInputChange}
-                multiline
-                editable={!isPostingComment} // Keep this, as posting new comment is separate
-              />
-              <TouchableOpacity
-                style={[styles.postCommentButton, isPostingComment || !commentInputText.trim() ? styles.postCommentButtonDisabled : {}]}
-                onPress={onPostComment}
-                disabled={isPostingComment || !commentInputText.trim()}
-              >
-                {isPostingComment ? (
-                  <ActivityIndicator size="small" color={colorScheme === 'dark' ? '#FFFFFF' : '#007AFF'} />
-                ) : (
-                  <ThemedText style={styles.postCommentButtonText}>Post</ThemedText>
-                )}
-              </TouchableOpacity>
-            </View>
-          )}
-        </View>
-      )}
-    </View>
-  );
-}
-
 
 export default function ExploreScreen() {
   const navigation = useNavigation();
