+++ mobile/app/(tabs)/index.tsx
@@ -17,6 +17,7 @@ import { ThemedText } from '@/components/ThemedText';
 import { useNavigation, useFocusEffect, useRoute } from '@react-navigation/native';
 import { AuthContext } from '../_layout';
 import { ScrollView } from 'react-native';
+import CheckBox from '../components/CheckBox';
 
 const HOST = Platform.select({ android: '10.0.2.2', ios: 'localhost' , web: 'localhost' });
 const API_BASE = `http://${HOST}:8080/api/auth`;
@@ -37,13 +38,7 @@ type TrendingPost = {
   photoUrl      : string | null;
 };
 
-function CheckBox({ checked, onPress }: { checked: boolean; onPress: () => void }) {
-  return (
-    <TouchableOpacity style={styles.checkbox} onPress={onPress}>
-      {checked && <Ionicons name="checkmark" size={16} color="#fff" />}
-    </TouchableOpacity>
-  );
-}
+
 
 export default function HomeScreen() {
   const navigation = useNavigation<Navigation>();
@@ -115,11 +110,9 @@ export default function HomeScreen() {
     (async () => {
       const token      = await AsyncStorage.getItem('token');
       const storedUser = await AsyncStorage.getItem('username');
-      const storedUserId = await AsyncStorage.getItem('user_id');
-      if (token && storedUser && storedUserId) {
+      if (token && storedUser) {
         setUserType('user');
         setUsername(storedUser);
-        setUserId(storedUserId);
         setLoggedIn(true);
         navigation.navigate('explore');
       }
@@ -183,20 +176,17 @@ export default function HomeScreen() {
       }
   
       // success path
-      const { token, username, user_id } = (await res.json()) as {
+      const { token, username } = (await res.json()) as {
         token: string;
 
         username: string;
-        user_id : string;
       };
       await AsyncStorage.multiSet([
         ['token', token],
-        ['username', username],
-        ['user_id' , user_id.toString()],
+        ['username', username]
       ]);
       setUserType('user');
       setUsername(username);
-      setUserId(user_id);
       setLoggedIn(true);
   
     } catch (error: any) {
@@ -567,14 +557,6 @@ const styles = StyleSheet.create({
     marginTop     : 4,
   },
   kvkkText: { marginLeft: 8, color: '#fff' },
-  checkbox: {
-    width: 20,
-    height: 20,
-    borderWidth: 1,
-    borderColor: '#fff',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
 
   errorBox: {
     position: 'absolute',
